Safari-611.1.10/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/mathml/presentation-markup/mrow/merror-001-expected.txt


FAIL Default CSS properties on merror assert_equals: expected "1px" but got "0px"
FAIL Bounding box is the same as mrow + 1px border assert_equals: expected 202 but got 133.34375
AN ERROR
AN ERROR
